@@ -1,7 +1,7 @@
 name: signalbutton
 description: push button signal simulator
 publish_to: 'none'
-version: 1.0.7+7
+version: 1.1.8+18
 
 environment:
   sdk: ">=2.16.2 <3.0.0"
@@ -10,26 +10,25 @@ dependencies:
   flutter:
     sdk: flutter
   # Firebase Analytics
-  firebase_core: ^1.24.0
-  firebase_analytics: ^9.3.8
+  firebase_core: ^2.9.0
+  firebase_analytics: ^10.2.0
   # App Icon
   cupertino_icons: ^1.0.5
-  flutter_launcher_icons: ^0.10.0
+  flutter_launcher_icons: ^0.12.0
   # Audio Player
-  audioplayers: ^1.1.1
+  audioplayers: ^3.0.1
   # Vibration
   vibration: ^1.7.6
   # Text to Speech
-  flutter_tts: ^3.5.3
+  flutter_tts: ^3.6.3
   # Settings
   flutter_settings_screens: ^0.3.3-null-safety+2
+  settings_ui: ^2.0.2
   # SharedPreference
-  shared_preferences: ^2.0.15
-  # Device Info
-  device_info_plus: ^5.0.5
+  shared_preferences: ^2.0.20
   # Admob
-  google_mobile_ads: ^2.0.1
-  app_tracking_transparency: ^2.0.2+4
+  google_mobile_ads: ^2.4.0
+  app_tracking_transparency: ^2.0.3
   # Localize
   flutter_localizations:
     sdk: flutter
@@ -37,24 +36,17 @@ dependencies:
   #env
   flutter_dotenv: ^5.0.2
   #Revenue Cat
-  purchases_flutter: ^4.3.1
+  purchases_flutter: ^4.11.1
   #MVVM
-  riverpod: ^2.1.1
-  flutter_riverpod: ^2.1.1
-  hooks_riverpod: ^2.1.1
-  flutter_hooks: ^0.18.5+1
-  freezed: ^2.2.1
-  freezed_annotation: ^2.2.0
-  build_runner: ^2.3.2
-  json_annotation: ^4.7.0
-  json_serializable: ^6.5.4
+  hooks_riverpod: ^2.3.2
+  flutter_hooks: ^0.18.6
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   flutter_lints: ^2.0.1
   # App Splash
-  flutter_native_splash: ^2.2.10+1
+  flutter_native_splash: ^2.2.19
 
 flutter_icons:
   image_path: "assets/images/icon.png"
@@ -81,7 +73,7 @@ flutter:
     - assets/images/traffic/uk/
     - assets/images/traffic/au/
     - assets/audios/
-    - .env
+    - assets/.env
 
   fonts:
     - family: freeTfb
