@@ -3,14 +3,16 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter_settings_screens/flutter_settings_screens.dart';
 import 'package:google_mobile_ads/google_mobile_ads.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'constant.dart';
 import 'firebase_options.dart';
-import 'settings.dart';
 import 'my_homepage.dart';
+import 'settings.dart';
+import 'upgrade.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -29,12 +31,15 @@ class MyApp extends StatelessWidget {
       localizationsDelegates: AppLocalizations.localizationsDelegates,
       supportedLocales: AppLocalizations.supportedLocales,
       title: 'LETS SIGNAL',
-      theme: ThemeData(primarySwatch: Colors.amber),
+      theme: ThemeData(
+        colorScheme: const ColorScheme.light(primary: greenColor),
+      ),
       debugShowCheckedModeBanner: false,
       home: const MyHomePage(),
       routes: {
         '/h' : (_) => const MyHomePage(),
         '/s' : (_) => const MySettingsPage(),
+        '/u' : (_) => const MyUpgradePage(),
       },
       navigatorObservers: <NavigatorObserver>[
         FirebaseAnalyticsObserver(analytics: FirebaseAnalytics.instance),
